meaningful commit message that describes the change you made to the file.
Here we generally write a message regarding what change we made to the file.Below the commit message fields, decide whether to add your commit to the current branch or to a new branch. If your current branch is master, you should choose to create a new branch for your commit and then create a pull request.